+++ src/dependency-map/index.js
@@ -1,3 +1,0 @@
-const DependencyMap = require('./dependency-map');
-
-module.exports = DependencyMap;