+++ src/exceptions/index.js
@@ -1,21 +1,11 @@
 import ComponentNotExistsException from './component-not-exists';
 import ComponentNotExistsInException from './component-not-exists-in';
-import VersionNotExistsException from './version-not-exists';
-import InvalidComponentIdException from './invalid-component-id';
-import InvalidBitJsonException from './invalid-bit-json';
 import NoConsumerFoundException from './no-consumer-found';
-import InvalidVersionException from './invalid-version';
 import MultipleScopesNoDefException from './multiple-scopes-no-def';
-import DuplicateComponentException from './duplicate-component';
 
 export {
   ComponentNotExistsException,
   ComponentNotExistsInException,
-  VersionNotExistsException,
-  InvalidComponentIdException,
-  InvalidBitJsonException,
   NoConsumerFoundException,
-  InvalidVersionException,
   MultipleScopesNoDefException,
-  DuplicateComponentException,
 };