+++ src/importer/component/index.js
@@ -1,3 +1,0 @@
-import Component from './component';
-
-export default Component;