+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Change
 
+- Move the writing-to-fs functionality to [bit-scope-client](https://github.com/teambit/bit-scope-client) project
 - Add opts to the load function
 - Add resolve function
 - Fix dist bug